% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_gates}
\alias{example_gates}
\title{example_gates}
\format{
A tibble
}
\usage{
data(example_gates)
}
\description{
Previously recorded brush path to select six region of space.
}
\keyword{datasets}
