% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transcriptomics.R
\docType{data}
\name{gate_seurat_obj}
\alias{gate_seurat_obj}
\title{gate_seurat_obj}
\format{
A list containing x, y coordinates for one gate.
}
\usage{
data(gate_seurat_obj)
}
\description{
Coordinates for a gate interactively drawn using \code{tidygate} on a Seurat object.
}
\details{
These x, y coordinates represent a manually defined gate used to subset specific populations
from a Seurat object.
}
\keyword{datasets}
