% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transcriptomics.R
\docType{data}
\name{pbmc_h3k4me3_hg38}
\alias{pbmc_h3k4me3_hg38}
\title{pbmc_h3k4me3_hg38}
\format{
A \code{GRanges} object with columns:
\describe{
\item{signalValue}{The signal value for each peak.}
\item{qValue}{The q-value for the significance of each peak.}
\item{peak}{The peak location in the genome.}
}
}
\source{
H3K4me3 peak data was lifted from hg19 to hg38 using UCSC liftOver chain files.
}
\usage{
data(pbmc_h3k4me3_hg38)
}
\value{
A GRanges object containing the H3K4me3 peaks for PBMCs in the hg38 genome assembly.
}
\description{
H3K4me3 peaks for human PBMCs lifted over to hg38 genome assembly
}
\details{
This object contains H3K4me3 peaks for human peripheral blood mononuclear cells (PBMCs) mapped to
the hg38 genome assembly. The peaks were initially from hg19 and were lifted over to hg38 using
a UCSC chain file. The peaks contain several columns of information including signal value,
q-value, and peak locations, which can be used for visualisation and analysis.

The H3K4me3 peaks were lifted over from the hg19 to hg38 genome assembly using the \code{liftOver}
function from the \code{rtracklayer} package. The \code{seqlevels} and \code{seqinfo} were aligned with the
\code{SingleCellExperiment} object \code{sce_sub}. The Ensembl-style chromosome naming convention (NCBI)
was applied for consistency. This dataset contains selected columns such as \code{signalValue},
\code{qValue}, and \code{peak}, making it ready for downstream analysis or plotting.
}
\examples{
# Load the dataset
data(pbmc_h3k4me3_hg38)

# Plot q-values of the H3K4me3 peaks
plot(pbmc_h3k4me3_hg38$qValue, type="l", ylab="q-value", main="H3K4me3 peaks")
abline(v=5000, lty=2)

}
\keyword{datasets}
