% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transcriptomics.R
\docType{data}
\name{pseudo_bulk}
\alias{pseudo_bulk}
\title{pseudo_bulk}
\format{
A \code{SummarizedExperiment} object.
}
\usage{
data(pseudo_bulk)
}
\description{
A \code{SummarizedExperiment} object containing pseudo-bulk RNA sequencing data.
}
\details{
This object was created for benchmarking purposes, representing a pseudo-bulk aggregation
of single-cell data, saved due to a specific issue with GitHub actions while functioning
correctly in local environments.
}
\keyword{datasets}
