% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transcriptomics.R
\docType{data}
\name{sce_obj}
\alias{sce_obj}
\title{sce_obj}
\format{
A \code{SingleCellExperiment} object.
}
\usage{
data(sce_obj)
}
\description{
A \code{SingleCellExperiment} dataset containing single-cell RNA sequencing data.
}
\details{
This SingleCellExperiment (SCE) object contains pre-processed single-cell RNA-seq data
ready for downstream analysis, such as clustering or differential expression analysis.
}
\keyword{datasets}
