% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transcriptomics.R
\docType{data}
\name{sce_obj_UMAP3}
\alias{sce_obj_UMAP3}
\title{sce_obj_UMAP3}
\format{
A \code{SingleCellExperiment} object.
}
\usage{
data(sce_obj_UMAP3)
}
\description{
A \code{SingleCellExperiment} dataset of single-cell RNA sequencing data with 3 UMAP dimensions.
}
\details{
This SCE object contains three-dimensional UMAP embeddings, which allow visualization
of cell populations in reduced-dimensional space.
}
\keyword{datasets}
