% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transcriptomics.R
\docType{data}
\name{seurat_obj}
\alias{seurat_obj}
\title{seurat_obj}
\format{
A \code{Seurat} object.
}
\usage{
data(seurat_obj)
}
\description{
A Seurat dataset containing single-cell RNA sequencing data.
}
\details{
This Seurat object contains pre-processed single-cell RNA-seq data that can be used
for downstream analysis such as clustering, differential expression, and visualization.
}
\keyword{datasets}
