% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transcriptomics.R
\docType{data}
\name{theme_multipanel}
\alias{theme_multipanel}
\title{theme_multipanel}
\format{
theme
}
\usage{
data(theme_multipanel)
}
\description{
A file holding a theme object.
}
\keyword{datasets}
