% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transcriptomics.R
\docType{data}
\name{tidygate_env_gates}
\alias{tidygate_env_gates}
\title{tidygate_env_gates}
\format{
A list containing the x and y coordinates for gates.
}
\source{
Created interactively using the \code{tidygate} package
}
\usage{
data(tidygate_env_gates)
}
\value{
A list of x and y coordinates representing the interactive gates.
}
\description{
A set of gates interactively drawn for spatial data using the \code{tidygate} package.
}
\details{
This object contains gates that were drawn interactively on spatial data based on tissue morphology,
highlighting areas of interest for future reproducible analysis. These gates can be loaded and applied
to similar data for consistent and automated gating.

Gates were saved programmatically for future use. The object was created using the \code{gate} function
in \code{tidygate} and saved as an RDS file for reproducibility. This can be particularly useful in spatial
transcriptomics data analysis where regions of interest are manually gated.
}
\examples{
# Load the gates
data(tidygate_env_gates)

# Apply the gates to new data for reproducible analysis
# spatial_data <- applyGates(spatial_data, gates = tidygate_env_gates)

}
\keyword{datasets}
