% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transcriptomics.R
\docType{data}
\name{theme_multipanel}
\alias{theme_multipanel}
\title{theme_multipanel}
\format{
A \code{ggplot2} theme.
}
\usage{
data(theme_multipanel)
}
\description{
A ggplot2 theme object designed for creating multipanel plots.
}
\details{
This theme can be applied to ggplot2-based visualizations to standardize the look and feel
of multipanel figures, ensuring consistency across plots.
}
\keyword{datasets}
